Abstracted out changing password and security question with Okta from /firstlogin endpoint.
I also renamed the endpoint /firstlogin.
